Week 2/requirements_pip.txt
absl-py==1.4.0
aiofiles==23.1.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argcomplete==2.1.2
asttokens==2.2.1
backcall==0.2.0
black==23.1.0
boto3==1.26.94
botocore==1.29.94
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
dacite==1.7.0
debugpy==1.6.6
decorator==5.1.1
Deprecated==1.2.13
-e git+https://github.com/facebookresearch/detectron2.git@072713649a9b0069c10aad1aaab819112e8f1e2e#egg=detectron2
dill==0.3.6
dnspython==2.3.0
eventlet==0.33.3
executing==1.2.0
fiftyone==0.19.1
fiftyone-brain==0.10.0
fiftyone-db==0.4.0
fonttools==4.39.2
future==0.18.3
fvcore==0.1.5.post20221221
glob2==0.7
google-auth==2.16.2
google-auth-oauthlib==0.4.6
graphql-core==3.2.3
greenlet==2.0.2
grpcio==1.51.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hydra-core==1.3.2
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
iopath==0.1.9
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
kaleido==0.2.1
kiwisolver==1.4.4
lazy_loader==0.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mongoengine==0.24.2
motor==3.1.1
mypy-extensions==1.0.0
ndjson==0.3.1
nest-asyncio==1.5.6
networkx==3.0
numpy @ file:///D:/bld/numpy_1675642778044/work
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.11.1
patool==1.12
pickleshare==0.7.5
Pillow @ file:///D:/bld/pillow_1675487347122/work
platformdirs==3.1.1
plotly==5.13.1
portalocker==2.7.0
pprintpp==0.4.0
priority==2.0.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
Pygments==2.14.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==305
PyYAML==6.0
pyzmq==25.0.1
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sse-starlette==0.10.3
sseclient-py==1.7.2
stack-data==0.6.2
starlette==0.20.4
strawberry-graphql==0.138.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.3.15
toml==0.10.2
tomli==2.0.1
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
tzdata==2022.7
tzlocal==4.3
universal-analytics-python3==1.1.1
urllib3==1.26.15
voxel51-eta==0.8.4
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
wsproto==1.2.0
xmltodict==0.13.0
yacs==0.1.8
zipp==3.15.0
